tests/requirements.txt
pytest==8.3.4
httpx==0.28.1
python-dotenv==1.1.0
playwright==1.50.0
pytest-asyncio==0.25.3
pytest-playwright==0.6.2
